Chore : add typescript
