+++ frontend/package.json
@@ -29,8 +29,11 @@
     "vite-plugin-svgr": "^2.2.2"
   },
   "devDependencies": {
-    "@types/react": "^18.0.24",
-    "@types/react-dom": "^18.0.8",
+    "@types/jest": "^29.4.0",
+    "@types/node": "^18.11.18",
+    "@types/react": "^18.0.27",
+    "@types/react-dom": "^18.0.10",
+    "@types/styled-components": "^5.1.26",
     "@vitejs/plugin-react": "^2.2.0",
     "eslint": "^8.27.0",
     "eslint-config-prettier": "^8.5.0",
@@ -41,6 +44,7 @@
     "msw": "^0.48.1",
     "prettier": "^2.7.1",
     "stylelint": "^14.14.1",
+    "typescript": "^4.8.4",
     "vite": "^3.2.3",
     "vite-plugin-html-env": "^1.2.7"
   },
